requirements.txt
anyio==4.4.0
argon2-cffi==23.1.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.37.13
botocore==1.37.13
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.3
Django==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fastjsonschema==2.20.0
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==10.2.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.14.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.2.1
uvicorn==0.34.0
waitress==3.0.0
Werkzeug==3.0.1
celery
amqp
django-celery-beat
